Migrate downstream packages to DTypeTag, update docs, and make NDArray sealed with leaf DType getters

+++ pkgs/notebook/lib/src/kernel_helper.dart
@@ -992,8 +992,8 @@ Plot plotSymbolic(
   }
   final lambda = f.lambdify([varName]);
   final xArr = linspace<Float64>(
-    Float64(from.toDouble()),
-    Float64(to.toDouble()),
+    from.toDouble(),
+    to.toDouble(),
     points,
     dtype: DType.float64,
   );
